refactor submit jobs widgets/funcs
